+++ src/GameDesignWizard.Core/Ideas/MediaAttachmentType.cs
@@ -1,0 +1,7 @@
+namespace GameDesignWizard.Core.Ideas;
+
+public enum MediaAttachmentType
+{
+    Image,
+    Audio
+}